+++ src/cpp/session/SessionMain.cpp
@@ -1834,7 +1834,10 @@ int main(int argc, char * const argv[])
       // will get re-initialized below)
       std::string programId = "rsession-" + core::system::username();
       core::log::setProgramId(programId);
-      core::system::initializeLog(programId, core::log::LogLevel::WARN);
+      core::system::initializeLog(programId,
+                                  core::log::LogLevel::WARN,
+                                  core::system::xdg::userLogDir(),
+                                  true); // force log dir to be under user's home directory
 
       // ignore SIGPIPE
       Error error = core::system::ignoreSignal(core::system::SigPipe);
@@ -1953,7 +1956,8 @@ int main(int argc, char * const argv[])
          {
             core::system::initializeLog(options.programIdentity(),
                                         core::log::LogLevel::WARN,
-                                        options.userLogPath());
+                                        options.userLogPath(),
+                                        true); // force log dir to be under user's home directory
          }
       }