+++ src/cpp/shared_core/include/shared_core/Logger.hpp
@@ -73,6 +73,7 @@ struct RefreshParams
     *        changing process running user to ensure logs can still be written to.
     */
    boost::optional<core::system::User> newUser;
+   bool chownLogDir;
 #endif
 };